requirements.txt
streamlit>=1.49.1
pandas>=2.3.2
google-genai>=1.36.0
pillow>=11.3.0
plotly>=6.3.0
qrcode>=8.2
openai>=1.107.1
djaodjin-pages>=0.8.6
